+++ app/controllers/ej2_showcase_controller.rb
@@ -1,4 +1,6 @@
 class Ej2ShowcaseController < ApplicationController
+  allow_unauthenticated_access
+
   def show
     @active_section = params[:section] || "forms"
   end